chore: Parse previousOfferingLetter attribute in CpMemoBandingService
This commit adds conditional parsing of the `previousOfferingLetter` attribute in the `CpMemoBandingService` class. If the attribute is a string, it is parsed and the parsed value is assigned to the corresponding key in the `parsed` object. If the attribute is not a string, the value is directly assigned to the `parsed` object. This change improves the handling of the `previousOfferingLetter` attribute in the service.